Move daemon file preparation to `DaemonInstance`
